Complete Rust 1.93 strict lint compatibility

+++ crates/dasobjectstore-gui-api/src/home_aggregator.rs
@@ -340,8 +340,8 @@ pub(crate) fn capacity_for_root(path: &Path) -> Option<FilesystemCapacity> {
     let stat = unsafe { stat.assume_init() };
     let fragment_size = stat.f_frsize;
     Some(FilesystemCapacity {
-        total_bytes: (stat.f_blocks as u64).saturating_mul(fragment_size),
-        available_bytes: (stat.f_bavail as u64).saturating_mul(fragment_size),
+        total_bytes: stat.f_blocks.saturating_mul(fragment_size),
+        available_bytes: stat.f_bavail.saturating_mul(fragment_size),
     })
 }
